pyproject.toml
[project]
name = "outreach-campaign"
version = "0.1.0"
description = "Multi-channel outreach campaign manager for crypto fund allocators"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12",
    "jinja2>=3.1",
    "thefuzz[speedup]>=0.22",
    "python-levenshtein>=0.25",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "markdown2>=2.4",
    "premailer>=3.10",
    "httpx>=0.27",
    "rich>=13.0",
    "tabulate>=0.9",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.9",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.150",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "testing.postgresql>=1.3",
]

[project.scripts]
outreach = "src.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
